Fix cache srm (#1494)

+++ functions/ToolScripts/emuDeckSRM.sh
@@ -192,9 +192,10 @@ SRM_deleteCache(){
     STEAMPATH="${HOME}/.steam/steam"
   else
     echo "Steam install not found"
+    return
   fi
 
-		zenity --question \
+  zenity --question \
     --text="If you are experiencing freezing or crashing with Steam ROM Manager, this will delete the cache and reset your non-Steam library. \
     \nThis will delete all of your non-Steam shortcuts. \
     \nDo note this includes any game launchers, browsers, fan games, any games or applications you have added as a non-Steam game. This will not delete the games or applications themselves. \
@@ -207,25 +208,25 @@ SRM_deleteCache(){
     \n \
     \nWould you like to delete the cache?" \
     --title="Delete Cache" \
-		--width=400 \
-		--height=300 \
-    --ok-label="No" \
-    --cancel-label="Yes"
+    --width=400 \
+    --height=300 \
+    --ok-label="Yes" \
+    --cancel-label="No"
 
+  response=$?
 
-		if [ $? = 1 ]; then
-      find "$STEAMPATH/userdata" -mindepth 2 -maxdepth 2 -type d -name 'config' -exec rm -rf {} +
-      echo "Cache deleted."
-      zenity --info \
+  # Only proceed if the user explicitly confirmed (pressed "Yes")
+  if [ "$response" -eq 0 ]; then
+    find "$STEAMPATH/userdata" -mindepth 2 -maxdepth 2 -type d -name 'config' -exec rm -rf {} +
+    echo "Cache deleted successfully."
+    zenity --info \
       --text="The cache has been deleted. All of your non-Steam shortcuts have been wiped. You may open Steam ROM Manager and re-add your ROMs to your library." \
       --title="Cache deleted" \
       --width=400 \
       --height=300
-    else
-      echo "User declined deleting cache."
-    fi
-
+  fi
 }
+
 
 SRM_addExtraParsers(){
   rsync -avhp --mkpath "$emudeckBackend/configs/steam-rom-manager/userData/userConfigurations.json" "$HOME/.config/steam-rom-manager/userData/" --backup --suffix=.bak
